+++ packages/headless/rollup.config.js
@@ -7,14 +7,23 @@ import {sizeSnapshot} from 'rollup-plugin-size-snapshot';
 import alias from '@rollup/plugin-alias';
 import {resolve as pathResolve} from 'path';
 import dts from "rollup-plugin-dts";
+import {readFileSync} from 'fs';
 
 const typescript = () => tsPlugin({tsconfig: './src/tsconfig.build.json'});
 const isCI = process.env.CI === 'true';
 const isProduction = process.env.BUILD === 'production';
 
+/**
+ * @returns {string}
+ */
+function getPackageVersion() {
+  return JSON.parse(readFileSync('package.json', 'utf-8')).version;
+}
+
 function replace() {
   const env = isProduction ? 'production' : 'development';
-  return replacePlugin({'process.env.NODE_ENV': JSON.stringify(env)});
+  const version = getPackageVersion();
+  return replacePlugin({'process.env.NODE_ENV': JSON.stringify(env), 'process.env.VERSION': JSON.stringify(version)});
 }
 
 function onWarn(warning, warn) {