+++ packages/headless/src/app/search-app-reducers.ts
@@ -27,6 +27,7 @@ import {debugReducer} from '../features/debug/debug-slice';
 import {facetOrderReducer} from '../features/facets/facet-order/facet-order-slice';
 import {redo, snapshot, undo} from '../features/history/history-actions';
 import {resultPreviewReducer} from '../features/result-preview/result-preview-slice';
+import {versionReducer} from '../features/debug/version-slice';
 
 /**
  * Map of reducers that make up the SearchAppState.
@@ -64,4 +65,5 @@ export const searchAppReducers: ReducersMapObject<SearchAppState> = {
   searchHub: searchHubReducer,
   debug: debugReducer,
   resultPreview: resultPreviewReducer,
+  version: versionReducer,
 };