+++ packages/headless/src/features/facets/category-facet-set/category-facet-reducer-helpers.test.ts
@@ -4,20 +4,20 @@ import {selectPath} from './category-facet-reducer-helpers';
 
 describe('category facet reducer helpers', () => {
   describe('#selectPath', () => {
-    const retrieveCount = 10;
+    const initialNumberOfValues = 10;
 
     describe('when the path is populated array', () => {
       it('sets #currentValues to the expected nested value', () => {
         const request = buildMockCategoryFacetRequest();
         const path = ['a', 'b'];
 
-        selectPath(request, path, retrieveCount);
+        selectPath(request, path, initialNumberOfValues);
 
         const b = buildMockCategoryFacetValueRequest({
           value: 'b',
           state: 'selected',
           retrieveChildren: true,
-          retrieveCount,
+          retrieveCount: initialNumberOfValues,
         });
 
         const a = buildMockCategoryFacetValueRequest({
@@ -25,7 +25,7 @@ describe('category facet reducer helpers', () => {
           children: [b],
           retrieveChildren: false,
           state: 'idle',
-          retrieveCount,
+          retrieveCount: initialNumberOfValues,
         });
 
         expect(request.currentValues).toEqual([a]);
@@ -33,7 +33,7 @@ describe('category facet reducer helpers', () => {
 
       it('sets #numberOfValues to 1', () => {
         const request = buildMockCategoryFacetRequest();
-        selectPath(request, ['a'], retrieveCount);
+        selectPath(request, ['a'], initialNumberOfValues);
 
         expect(request.numberOfValues).toBe(1);
       });
@@ -46,23 +46,23 @@ describe('category facet reducer helpers', () => {
         ];
         const request = buildMockCategoryFacetRequest({currentValues});
 
-        selectPath(request, [], retrieveCount);
+        selectPath(request, [], initialNumberOfValues);
 
         expect(request.currentValues).toEqual([]);
       });
 
-      it('does not adjust the #numberOfValues of the request', () => {
+      it('resets the number of values to the initialNumberOfValues', () => {
         const numberOfValues = 5;
         const request = buildMockCategoryFacetRequest({numberOfValues});
-        selectPath(request, [], retrieveCount);
+        selectPath(request, [], initialNumberOfValues);
 
-        expect(request.numberOfValues).toBe(numberOfValues);
+        expect(request.numberOfValues).toBe(initialNumberOfValues);
       });
     });
 
     it('sets the request #preventAutoSelect to true', () => {
       const request = buildMockCategoryFacetRequest({preventAutoSelect: false});
-      selectPath(request, [], retrieveCount);
+      selectPath(request, [], initialNumberOfValues);
 
       expect(request.preventAutoSelect).toBe(true);
     });