+++ packages/headless/src/features/facets/category-facet-set/category-facet-set-slice.ts
@@ -58,7 +58,7 @@ export const categoryFacetSetReducer = createReducer(
         Object.keys(state).forEach((id) => {
           const request = state[id]!.request;
           const path = cf[id] || [];
-          selectPath(request, path, request.numberOfValues);
+          selectPath(request, path, state[id]!.initialNumberOfValues);
         });
       })
       .addCase(updateCategoryFacetSortCriterion, (state, action) => {
@@ -135,15 +135,15 @@ export const categoryFacetSetReducer = createReducer(
         handleCategoryFacetNestedNumberOfValuesUpdate(state, action.payload);
       })
       .addCase(selectCategoryFacetSearchResult, (state, action) => {
-        const {facetId, value, retrieveCount} = action.payload;
-        const request = state[facetId]?.request;
+        const {facetId, value} = action.payload;
+        const facet = state[facetId];
 
-        if (!request) {
+        if (!facet) {
           return;
         }
 
         const path = [...value.path, value.rawValue];
-        selectPath(request, path, retrieveCount);
+        selectPath(facet.request, path, facet.initialNumberOfValues);
       })
       .addCase(executeSearch.fulfilled, (state, action) => {
         const {facets} = action.payload.response;